+++ public/app/features/explore/Logs/LogsSamplePanel.tsx
@@ -14,10 +14,10 @@ import {
 import { config, reportInteraction } from '@grafana/runtime';
 import { DataQuery, TimeZone } from '@grafana/schema';
 import { Button, Collapse, Icon, Tooltip, useStyles2 } from '@grafana/ui';
-import { dataFrameToLogsModel } from 'app/core/logsModel';
 import store from 'app/core/store';
 
 import { LogRows } from '../../logs/components/LogRows';
+import { dataFrameToLogsModel } from '../../logs/logsModel';
 import { SupplementaryResultError } from '../SupplementaryResultError';
 
 import { SETTINGS_KEYS } from './utils/logs';
